requirements/service.txt
fastapi==0.121.0
matplotlib>=3.8,<3.11
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
pillow==12.0.0
pydantic==2.12.3
python-multipart==0.0.22
PyYAML==6.0.3
uvicorn==0.42.0
